% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Get metadata for a single dataset}
\usage{
get_dataset(id)
}
\arguments{
\item{id}{Dataset id from \code{\link[=list_datasets]{list_datasets()}}.}
}
\value{
A list with \code{name}, \code{description}, \code{citation}.
}
\description{
Get metadata for a single dataset
}
