% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iearthdata-package.R
\docType{package}
\name{iearthdata-package}
\alias{iearthdata}
\alias{iearthdata-package}
\title{iearthdata: Client for the iEarth DataHub API}
\description{
An R client for the iEarth DataHub at data-starcloud.pcl.ac.cn. Performs bounding-box and time-period queries against hosted Earth- observation datasets and returns pre-signed download URLs for the intersecting assets. Primary use case is access to the SDC30 Embedded Seamless Data (ESD) global land monitoring dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shuangchencc/ESD}
}

}
\author{
\strong{Maintainer}: Hugh \email{hugh@belian.earth}

Authors:
\itemize{
  \item Hugh \email{hugh@belian.earth}
}

}
\keyword{internal}
